Added the LockedAt feature + Started designing Ticket Service and Controller

+++ src/main/java/com/example/bookmyshow_v1/models/MovieShow.java
@@ -19,6 +19,6 @@ public class MovieShow extends BaseModel{
     private Auditorium auditorium;
     private Date entryTime;
     private Date exitTime;
-    private int lockingTimeout;
+    private long lockingTimeout;
     private int maxTicketsAllowedForOneUser;
 }
